Extract rule group processors from formatQuery.ts
